// portfolio_next/app/types/project.ts
export interface ProjetoUIDesign {
  id: string;
  name: string;
  description: string;
  capaUrl: string;
  typeOfCase?: string;
  typeOfProject?: string;
  behanceLink?: string;
}

export interface ProjetoFrontEnd {
  id: string;
  name: string;
  description: string;
  capaUrl: string;
  typeOfCase?: string;
  typeOfProject?: string;
  GitHubLink?: string;
  HostLink?: string;
}